pyproject.toml
[project]
name = "python-scripts"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "FrankHuang", email = "pingrunhuang@gmail.com"},
]
dependencies = [
    "openpyxl>=3.1.5",
    "requests>=2.32.3",
    "colorama>=0.4.6",
    "pyyaml>=6.0.1",
    "python-binance>=1.0.19",
]
requires-python = "==3.9.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"
